WIP: Move TMC implementation to its own repo. Dependencies ok Does not compile yet: Missing a cached informerfactory for TMC resources to be passed by
This is plain portage, no attempt to improve design or code structure to reflect that TMC is not part of kcp core anymore.
Wondering whether the approach of creating a new server extending the "core server" instead of using the plain APIExport mechanism is necessary. Is it for kcp clustering?
Would it make sense to have in TMC a separate module (sdk) for the API and client like in kcp?

Signed-off-by: Frederic Giloux <fgiloux@redhat.com>